CORDA-265: Refactor construction of dummy RPC client into an abstract base class.
